requirements.txt
# 后端框架
fastapi==0.115.0
uvicorn==0.30.6
# 向量数据库
chromadb==0.5.15
# RAG相关工具链
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-huggingface==0.1.0
# 接口与文件处理
python-multipart==0.0.12
python-dotenv==1.0.1
# 前端框架
#streamlit==1.38.0
# 网络请求
requests==2.32.3
# 重排序优化（复试核心亮点）
cohere==5.11.0